requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-ckeditor==5.7.0
django-js-asset==1.2.2
django-mailbox==4.8.0
django-phonenumber-field==3.0.1
django-project==0.7
django-webpack-loader==0.6.0
django-xhtml2pdf==0.0.4
email==6.0.0a1
gunicorn==19.9.0
html5lib==1.0.1
humanize==0.5.1
phonenumbers==8.10.17
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2==2.8.3
PyPDF2==1.26.0
pytz==2019.1
reportlab==3.5.28
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
